started on exercise3
