+++ app/src/main/java/dk/itu/moapd/scootersharing/rasni/Scooter.kt
@@ -1,7 +1,11 @@
 package dk.itu.moapd.scootersharing.rasni
 
-    data class Scooter(var name: String, var location: String) {
+    data class Scooter(var name: String, var location: String, var timestamp: Long = System.currentTimeMillis()) {
         override fun toString(): String {
             return "[Scooter] $name is placed at $location ."
         }
+        fun getTimestamp(): String {
+            return timestamp.toString()
+        }
+
     }
