+++ app/src/main/res/layout/activity_main.xml
@@ -23,52 +23,24 @@
         android:focusableInTouchMode="true"
         android:gravity="center">
 
-        <com.google.android.material.textfield.TextInputLayout
-
+        <Button
+            android:id="@+id/start_ride"
+            style="@style/Widget.AppCompat.Button"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginBottom="10dp"
-            app:endIconMode="clear_text"
-            style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox">
-
-            <com.google.android.material.textfield.TextInputEditText
-                android:id="@+id/ScooterName"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableStart="@drawable/baseline_electric_scooter_24"
-                android:drawablePadding="10dp"
-                android:drawableTint="@color/purple_700"
-                android:hint="Enter scooter name"
-                android:selectAllOnFocus="false"
-                android:singleLine="false" />
-        </com.google.android.material.textfield.TextInputLayout>
-
-        <com.google.android.material.textfield.TextInputLayout
-
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="10dp"
-            app:endIconMode="clear_text"
-            style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox">
-
-            <com.google.android.material.textfield.TextInputEditText
-                android:id="@+id/ScooterLocation"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableStart="@drawable/baseline_location_on_24"
-                android:drawablePadding="10dp"
-                android:drawableTint="@color/purple_700"
-                android:hint="Enter scooter location" />
-        </com.google.android.material.textfield.TextInputLayout>
+            android:background="@drawable/button"
+            android:padding="10dp"
+            android:text="@string/start_ride_button" />
 
         <Button
-            android:id="@+id/Button"
+            android:id="@+id/update_ride"
             style="@style/Widget.AppCompat.Button"
-            android:layout_width="wrap_content"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:padding="10dp"
             android:background="@drawable/button"
-            android:text="Start ride" />
+            android:padding="10dp"
+            android:text="@string/start_update_button" />
+
 
     </LinearLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
