+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,7 @@
 <resources>
     <string name="app_name">Scooter Sharing</string>
+    <string name="start_ride_button">Start Ride</string>
+    <string name="start_update_button">Update Ride</string>
+    <string name="scooter_name_field">Enter scooter name</string>
+    <string name="scooter_location_field">Enter scooter location</string>
 </resources>
